fix: refactor to simplify
